analyzer/requirements.txt
psycopg2-binary==2.9.9
nats-py==2.6.0
pyyaml==6.0.1
jsonschema==4.19.2
aiohttp==3.9.1
python-dotenv==1.0.0
uvloop==0.19.0
